requirements.txt
face_recognition
fastapi
uvicorn
pillow
python-multipart
